+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Release 1.3.2 (2017-03-03)
 ==========================
 
 * Added new Pi models to stop :func:`pi_info` breaking
+* Fix issue with :func:`pi_info` breaking on unknown Pi models
 
 Release 1.3.1 (2016-08-31 ... later)
 ====================================